requirements.txt
robotframework>=7.3.2
pymongo>=4.15.4
python-dotenv>=1.2.1
boto3>=1.34.0
requests>=2.31.0
